pyproject.toml
[project]
dependencies = [
  "ipykernel>=6.29.5",
  "lightning>=2.5.2",
  "livelossplot>=0.5.6",
  "m3logpy>=0.1.0",
  "matplotlib>=3.10.3",
  "numpy>=2.3.1",
  "pydantic>=2.11.7",
  "pynvml>=12.0.0",
  "pyyaml>=6.0.2",
  "scikit-learn>=1.7.0",
  "scipy>=1.16.0",
  "seaborn>=0.13.2",
  "torch>=2.7.1",
  "torchaudio>=2.7.1",
  "torchvision>=0.22.1",
]
description = "Deep Learning Study Project"
name = "deeplearning_study" 
readme = "README.md"
requires-python = ">=3.13"
version = "0.1.0"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.uv.sources]
torch = [{index = "pytorch-cu128"}]
torchaudio = [{index = "pytorch-cu128"}]
torchvision = [{index = "pytorch-cu128"}]

[[tool.uv.index]]
explicit = true
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"

[dependency-groups]
dev = [
  "pyrefly>=0.21.0",
  "ruff>=0.12.0",
]

[tool.ruff]
include = ["**/*.py", "**/*.ipynb"]

[tool.pyrefly]
project-includes = ["**/*.py", "**/*.ipynb"]
search-paths = ["src"]
